remove duplicated logic
