+++ google-cloud-storage/src/test/java/com/google/cloud/storage/ITOpenTelemetryTestbenchTest.java
@@ -18,18 +18,17 @@ package com.google.cloud.storage;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-import com.google.cloud.NoCredentials;
 import com.google.cloud.ReadChannel;
 import com.google.cloud.WriteChannel;
 import com.google.cloud.storage.Storage.BlobSourceOption;
 import com.google.cloud.storage.Storage.BlobTargetOption;
 import com.google.cloud.storage.Storage.CopyRequest;
+import com.google.cloud.storage.TransportCompatibility.Transport;
 import com.google.cloud.storage.it.runner.StorageITRunner;
 import com.google.cloud.storage.it.runner.annotations.Backend;
+import com.google.cloud.storage.it.runner.annotations.CrossRun;
 import com.google.cloud.storage.it.runner.annotations.Inject;
-import com.google.cloud.storage.it.runner.annotations.SingleBackend;
 import com.google.cloud.storage.it.runner.registry.Generator;
-import com.google.cloud.storage.it.runner.registry.TestBench;
 import com.google.cloud.storage.otel.TestExporter;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
@@ -37,31 +36,31 @@ import io.opentelemetry.sdk.trace.SdkTracerProvider;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.export.SimpleSpanProcessor;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
-import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
-import java.util.UUID;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 @RunWith(StorageITRunner.class)
-@SingleBackend(Backend.TEST_BENCH)
-public class ITHttpOpenTelemetryTest {
-  @Inject public TestBench testBench;
-  private StorageOptions options;
-  private SpanExporter exporter;
-  private BlobId blobId;
-  private Storage storage;
-  private static final byte[] helloWorldTextBytes = "hello world".getBytes();
-  private static final byte[] helloWorldGzipBytes = TestUtils.gzipBytes(helloWorldTextBytes);
+@CrossRun(
+    backends = Backend.TEST_BENCH,
+    transports = {Transport.HTTP, Transport.GRPC})
+public class ITOpenTelemetryTestbenchTest {
+  @Inject public Transport transport;
   @Inject public Generator generator;
   @Inject public BucketInfo testBucket;
+  @Inject public Storage storage;
+  private StorageOptions options;
+  private SpanExporter exporter;
+  private static final byte[] helloWorldTextBytes = "hello world".getBytes();
+  private BlobId blobId;
+  private static final Path tmpDir = Paths.get(System.getProperty("java.io.tmpdir"));
 
   @Before
   public void setUp() {
@@ -73,13 +72,7 @@ public class ITHttpOpenTelemetryTest {
                     .addSpanProcessor(SimpleSpanProcessor.create(exporter))
                     .build())
             .build();
-    options =
-        StorageOptions.http()
-            .setHost(testBench.getBaseUri())
-            .setProjectId("projectId")
-            .setCredentials(NoCredentials.getInstance())
-            .setOpenTelemetrySdk(openTelemetrySdk)
-            .build();
+    options = storage.getOptions().toBuilder().setOpenTelemetrySdk(openTelemetrySdk).build();
     storage = options.getService();
     String objectString = generator.randomObjectName();
     blobId = BlobId.of(testBucket.getName(), objectString);
@@ -87,7 +80,7 @@ public class ITHttpOpenTelemetryTest {
 
   @Test
   public void runCreateBucket() {
-    String bucket = "random-bucket" + UUID.randomUUID();
+    String bucket = "random-bucket";
     storage.create(BucketInfo.of(bucket));
     TestExporter testExported = (TestExporter) exporter;
     List<SpanData> spanData = testExported.getExportedSpans();
@@ -102,49 +95,10 @@ public class ITHttpOpenTelemetryTest {
     TestExporter testExported = (TestExporter) exporter;
     List<SpanData> spanData = testExported.getExportedSpans();
     checkCommonAttributes(spanData);
-  }
-
-  @Test
-  public void runRead() throws IOException {
-    BlobInfo blobInfo =
-        BlobInfo.newBuilder(blobId).setContentEncoding("gzip").setContentType("text/plain").build();
-    storage.create(blobInfo, helloWorldGzipBytes);
-    Path helloWorldTxtGz = File.createTempFile(blobId.getName(), ".txt.gz").toPath();
-    storage.downloadTo(
-        blobId, helloWorldTxtGz, Storage.BlobSourceOption.shouldReturnRawInputStream(true));
-    TestExporter testExported = (TestExporter) exporter;
-    List<SpanData> spanData = testExported.getExportedSpans();
-    checkCommonAttributes(spanData);
-    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("read")));
-  }
-
-  @Test
-  public void runCopy() {
-
-    byte[] expected = "Hello, World!".getBytes(StandardCharsets.UTF_8);
-
-    BlobInfo info =
-        BlobInfo.newBuilder(testBucket.getName(), generator.randomObjectName() + "copy/src")
-            .build();
-    Blob cpySrc = storage.create(info, expected, BlobTargetOption.doesNotExist());
-
-    BlobInfo dst =
-        BlobInfo.newBuilder(testBucket.getName(), generator.randomObjectName() + "copy/dst")
-            .build();
-
-    CopyRequest copyRequest =
-        CopyRequest.newBuilder()
-            .setSource(cpySrc.getBlobId())
-            .setSourceOptions(BlobSourceOption.generationMatch(cpySrc.getGeneration()))
-            .setTarget(dst, BlobTargetOption.doesNotExist())
-            .build();
-    CopyWriter copyWriter = storage.copy(copyRequest);
-    copyWriter.getResult();
-    TestExporter testExported = (TestExporter) exporter;
-    List<SpanData> spanData = testExported.getExportedSpans();
-    checkCommonAttributes(spanData);
-    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("openRewrite")));
-    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("rewrite")));
+    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("create")));
+    Assert.assertTrue(
+        spanData.stream().anyMatch(x -> x.getName().contains("internalDirectUpload")));
+    Assert.assertEquals(spanData.get(1).getSpanContext(), spanData.get(0).getParentSpanContext());
   }
 
   @Test
@@ -155,27 +109,63 @@ public class ITHttpOpenTelemetryTest {
     TestExporter testExported = (TestExporter) exporter;
     List<SpanData> spanData = testExported.getExportedSpans();
     checkCommonAttributes(spanData);
-    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("load")));
+    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("readAllBytes")));
   }
 
   @Test
-  public void runCreateFromPath() throws IOException {
+  public void runCreateFrom() throws IOException {
     Path helloWorldTxtGz = File.createTempFile(blobId.getName(), ".txt.gz").toPath();
     storage.createFrom(BlobInfo.newBuilder(blobId).build(), helloWorldTxtGz);
     TestExporter testExported = (TestExporter) exporter;
     List<SpanData> spanData = testExported.getExportedSpans();
     checkCommonAttributes(spanData);
     Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("createFrom")));
+    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("internalCreateFrom")));
   }
 
   @Test
-  public void runCreateFromInputStream() throws IOException {
-    InputStream inputStream = new ByteArrayInputStream(helloWorldTextBytes);
-    storage.createFrom(BlobInfo.newBuilder(blobId).build(), inputStream);
+  public void runDownloadToPath() throws IOException {
+    BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType("text/plain").build();
+    storage.create(blobInfo, helloWorldTextBytes);
+    try (TmpFile file = TmpFile.of(tmpDir, "download-to", ".txt")) {
+      storage.downloadTo(blobId, file.getPath());
+      TestExporter testExported = (TestExporter) exporter;
+      List<SpanData> spanData = testExported.getExportedSpans();
+      checkCommonAttributes(spanData);
+      Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("downloadTo")));
+    }
+  }
+
+  @Test
+  public void runDownloadToOutputStream() {
+    BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType("text/plain").build();
+    storage.create(blobInfo, helloWorldTextBytes);
+    ByteArrayOutputStream baos = new ByteArrayOutputStream();
+    storage.downloadTo(blobId, baos);
     TestExporter testExported = (TestExporter) exporter;
     List<SpanData> spanData = testExported.getExportedSpans();
     checkCommonAttributes(spanData);
-    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("createFrom")));
+    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("downloadTo")));
+  }
+
+  @Test
+  public void runCopy() {
+    BlobInfo info =
+        BlobInfo.newBuilder(testBucket, generator.randomObjectName() + "copy/src").build();
+    Blob cpySrc = storage.create(info, helloWorldTextBytes, BlobTargetOption.doesNotExist());
+    BlobInfo dst =
+        BlobInfo.newBuilder(testBucket, generator.randomObjectName() + "copy/dst").build();
+    CopyRequest copyRequest =
+        CopyRequest.newBuilder()
+            .setSource(cpySrc.getBlobId())
+            .setSourceOptions(BlobSourceOption.generationMatch(cpySrc.getGeneration()))
+            .setTarget(dst, BlobTargetOption.doesNotExist())
+            .build();
+    storage.copy(copyRequest);
+    TestExporter testExported = (TestExporter) exporter;
+    List<SpanData> spanData = testExported.getExportedSpans();
+    checkCommonAttributes(spanData);
+    Assert.assertTrue(spanData.stream().anyMatch(x -> x.getName().contains("copy")));
   }
 
   @Test
@@ -209,7 +199,7 @@ public class ITHttpOpenTelemetryTest {
       Assert.assertEquals("googleapis/java-storage", getAttributeValue(span, "gcp.client.repo"));
       Assert.assertEquals(
           "com.google.cloud:google-cloud-storage", getAttributeValue(span, "gcp.client.artifact"));
-      Assert.assertEquals("http", getAttributeValue(span, "rpc.system"));
+      Assert.assertEquals(transport.name().toLowerCase(), getAttributeValue(span, "rpc.system"));
     }
   }
